requirements.txt
discord
pydub
python-magic
pynacl